+++ projects/angular/components/ui-grid/src/ui-grid.module.ts
@@ -13,7 +13,6 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { UiAutoAccessibleLabelModule } from '@uipath/angular/a11y';
 import { UiSuggestModule } from '@uipath/angular/components/ui-suggest';
-import { UiCustomMatMenuTriggerModule } from '@uipath/angular/directives/custom-mat-menu-trigger';
 import { UiNgLetModule } from '@uipath/angular/directives/ui-ng-let';
 import { UiVirtualScrollViewportResizeModule } from '@uipath/angular/directives/ui-virtual-scroll-viewport-resize';
 
@@ -55,7 +54,6 @@ import { UiGridComponent } from './ui-grid.component';
         UiVirtualScrollViewportResizeModule,
         UiAutoAccessibleLabelModule,
         UiNgLetModule,
-        UiCustomMatMenuTriggerModule,
     ],
     declarations: [
         UiGridComponent,