+++ projects/angular/components/ui-suggest/src/models/maxSelectionConfig.ts
@@ -1,0 +1,13 @@
+/**
+ * Suggest max selection config type definition.
+ *  count: Max number of selected items.
+ *  itemTooltip: Tooltip to be displayed on each unselected item.
+ *  footerMessage: Message to be displayed in the dropdown footer.
+ *
+ * @export
+ */
+export type SuggestMaxSelectionConfig = {
+    count: number;
+    itemTooltip: string;
+    footerMessage: string;
+};